backend/requirements/base.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# Celery
celery[redis]>=5.3,<6.0
django-celery-beat>=2.5,<3.0

# Authentication
PyJWT>=2.8,<3.0

# Encryption
cryptography>=41.0,<43.0

# LLM Integration
httpx>=0.26,<1.0

# HTTP Client
requests>=2.31,<3.0

# Utilities
python-dotenv>=1.0,<2.0
